Group events by faculty or room
